#282 feat: jsoneditor dynamically loaded
