+++ src/Template/Layout/js/app/directives/datepicker.js
@@ -3,6 +3,7 @@
  *
  */
 
+
 export default {
     install(Vue) {
         Vue.directive('datepicker', {
